refactor: remove module_iterate
